+++ lib/typus/orm/active_record/admin_user_v1.rb
@@ -44,11 +44,6 @@ module Typus
 
         module InstanceMethods
 
-          def to_label
-            full_name = [first_name, last_name].delete_if { |s| s.blank? }
-            full_name.any? ? full_name.join(" ") : email
-          end
-
           def locale
             (preferences && preferences[:locale]) ? preferences[:locale] : ::I18n.default_locale
           end