+++ test/app/models/typus_user_test.rb
@@ -107,15 +107,6 @@ class TypusUserTest < ActiveSupport::TestCase
   test "to_label" do
     user = FactoryGirl.build(:typus_user)
     assert_equal user.email, user.to_label
-
-    user = FactoryGirl.build(:typus_user, :first_name => "John")
-    assert_equal "John", user.to_label
-
-    user = FactoryGirl.build(:typus_user, :last_name => "Locke")
-    assert_equal "Locke", user.to_label
-
-    user = FactoryGirl.build(:typus_user, :first_name => "John", :last_name => "Locke")
-    assert_equal "John Locke", user.to_label
   end
 
   test "admin gets a list of all applications" do